Use two-column format for ART_APEX_JARS components.
This is to unify ART_APEX_JARS format with the new 2-column
PRODUCT_BOOT_JARS format. The first component is the ART apex name.
This is a prerequisite change for moving core-icu4j to a separate
com.android.i18n apex.

Test: aosp_walleye-userdebug boots
Bug: 138994281
Change-Id: I7159260d1f46094857de52b21335defd352ca9b7
Merged-In: I7159260d1f46094857de52b21335defd352ca9b7