Implement Reference.refersTo()
Avoid Reference.get(), so that we

a) Don't block for the GC when weak ref access is disabled.
b) Don't preserve a referent just because we compared to it.

Strengthen ConcurrentCopying::Copy memory ordering guarantees to make
this possible. In the process, weaken the existing fence to a
constructor fence, which should sometimes be slightly cheaper than
a release fence, which will be implemented as a full fence on ARM.

While looking at forwarding reference memory ordering, removed a
long identified unnecessary seq_cst memory access in
IsNullOrMarkedHeapReference.

Test: boot AOSP, atest ReferenceTest in libcore
Bug: 172573708
Merged-In: Ie716df684be94006353a54226f85ea93a7254b6c
Change-Id: Ie716df684be94006353a54226f85ea93a7254b6c