Do not open dex files in CLC if we only need to get dexopt status
The verifying the class loader context when calling GetDexOptNeeded
we only need the dex locations and the checksums. Opening the full
dex files may lead to in memory extraction which is expensive and
unnecessary.

Add a special path in ClassLoaderContext::OpenDexFiles which will
extract the locations and the checksums from the apk instead of
actually opening the dex files.

We re-uses the same logic in OpenDexFiles in order to avoid
implementing the opening algorithm twice (which, given all the edge
cases is not trivial).

Bug: 169869944
Test: test-art-host
Merged-In: Ic327889677ce697cd60c5c688281515b932a2a76
Change-Id: Ic327889677ce697cd60c5c688281515b932a2a76

(cherry picked from commit 5f751b7497216e07489c7dd68b70bd17101422fa)

Change-Id: I3c6f4610c365996b94d9bbe3c773bc73437ee1a5